pyproject.toml
[project]
name = "fast-spectral-ensemble-clustering"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Stepan Ponomarev", email = "stponomarevw@gmail.com" }
]
dependencies = [
    "numpy>=2.1.3",
    "scipy>=1.14.1",
    "scikit-learn>=1.5.2",
    "networkx>=3.4.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "optuna>=4.1.0",
    "plotly>=5.24.1",
    "ucimlrepo>=0.0.7",
    "torchvision>=0.20.1",
    "mlflow>=2.18.0",
    "pytorch-lightning>=2.4.0",
    "optuna-integration[pytorch_lightning]>=4.1.0",
    "lightning>=2.4.0",
    "kaleido>=0.2.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fsec", "data", "clustering_algorithms", "benchmarks"]
